Break the v2 manifest across modules
