+++ crates/manifest/src/schema/v2.rs
@@ -1,15 +1,27 @@
 use anyhow::{Context, anyhow};
-use itertools::Itertools;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use spin_serde::{DependencyName, DependencyPackageName, FixedVersion, LowerSnakeId};
 pub use spin_serde::{KebabId, SnakeId};
-use std::path::PathBuf;
 
 pub use super::common::{ComponentBuildConfig, ComponentSource, Variable, WasiFilesMount};
-use super::json_schema;
+use super::{json_schema, kebab_or_snake_case, one_or_many};
 
 pub(crate) type Map<K, V> = indexmap::IndexMap<K, V>;
+
+mod app_details;
+mod component;
+mod dependency;
+mod target_env;
+mod trigger;
+
+pub use app_details::AppDetails;
+pub use component::{
+    Component, ComponentDependencies, ComponentProfileBuildOverride, ComponentProfileOverride,
+};
+pub use dependency::{ComponentDependency, InheritConfiguration, TriggerDependency};
+pub use target_env::TargetEnvironmentRef;
+pub use trigger::{ComponentSpec, OneOrManyComponentSpecs, Trigger, TriggerDependencies};
 
 /// App manifest
 #[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
@@ -74,932 +86,9 @@ impl AppManifest {
     }
 }
 
-/// App details
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
-#[serde(deny_unknown_fields)]
-pub struct AppDetails {
-    /// The name of the application.
-    ///
-    /// Example: `name = "my-app"`
-    pub name: String,
-    /// The application version. This should be a valid semver version.
-    ///
-    /// Example: `version = "1.0.0"`
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub version: String,
-    /// A human-readable description of the application.
-    ///
-    /// Example: `description = "App description"`
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub description: String,
-    /// The author(s) of the application.
-    ///
-    /// `authors = ["author@example.com"]`
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub authors: Vec<String>,
-    /// The Spin environments with which application components must be compatible
-    /// unless otherwise specified. Individual components may express different
-    /// requirements: these override the application-level default.
-    ///
-    /// Example: `targets = ["spin-up:3.3", "spinkube:0.4"]`
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub targets: Vec<TargetEnvironmentRef>,
-    /// Application-level settings for the trigger types used in the application.
-    /// The possible values are trigger type-specific.
-    ///
-    /// Example:
-    ///
-    /// ```ignore
-    /// [application.triggers.redis]
-    /// address = "redis://notifications.example.com:6379"
-    /// ```
-    ///
-    /// Learn more (Redis example): https://spinframework.dev/redis-trigger#setting-a-default-server
-    #[serde(rename = "trigger", default, skip_serializing_if = "Map::is_empty")]
-    #[schemars(schema_with = "json_schema::map_of_toml_tables")]
-    pub trigger_global_configs: Map<String, toml::Table>,
-    /// Settings for custom tools or plugins. Spin ignores this field.
-    #[serde(default, skip_serializing_if = "Map::is_empty")]
-    #[schemars(schema_with = "json_schema::map_of_toml_tables")]
-    pub tool: Map<String, toml::Table>,
-}
-
-/// Trigger configuration. A trigger maps an event of the trigger's type (e.g.
-/// an HTTP request on route `/shop`, a Redis message on channel `orders`) to
-/// a Spin component.
-///
-/// The trigger manifest contains additional fields which depend on the trigger
-/// type. For the `http` type, these additional fields are `route` (required) and
-/// `executor` (optional). For the `redis` type, the additional fields are
-/// `channel` (required) and `address` (optional). For other types, see the trigger
-/// documentation.
-///
-/// Learn more: https://spinframework.dev/http-trigger, https://spinframework.dev/redis-trigger
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct Trigger {
-    /// Optional identifier for the trigger.
-    ///
-    /// Example: `id = "trigger-id"`
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub id: String,
-    /// The component that Spin should run when the trigger occurs. For HTTP triggers,
-    /// this is the HTTP request handler for the trigger route. This is typically
-    /// the ID of an entry in the `[component]` table, although you can also write
-    /// the component out as the value of this field.
-    ///
-    /// Example: `component = "shop-handler"`
-    ///
-    /// Learn more: https://spinframework.dev/triggers#triggers-and-components
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub component: Option<ComponentSpec>,
-    /// Additional components used when the trigger occurs.
-    /// The meaning of entries in this table is trigger-specific.
-    ///
-    /// `components = { ... }`
-    #[serde(default, skip_serializing_if = "Map::is_empty")]
-    pub components: Map<String, OneOrManyComponentSpecs>,
-    /// Additional components to be invoked during trigger processing.
-    /// The meaning of entries in this table is trigger-specific.
-    ///
-    /// `dependencies = { ... }`
-    #[serde(default, skip_serializing_if = "Map::is_empty")]
-    pub dependencies: Map<String, TriggerDependencies>,
-    /// Opaque trigger-type-specific config
-    #[serde(flatten)]
-    pub config: toml::Table,
-}
-
-/// One or many `ComponentSpec`(s)
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
-#[serde(transparent)]
-pub struct OneOrManyComponentSpecs(
-    #[serde(with = "one_or_many")]
-    #[schemars(schema_with = "json_schema::one_or_many::<ComponentSpec>")]
-    pub Vec<ComponentSpec>,
-);
-
-/// One or many `ComponentSpec`(s)
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
-#[serde(transparent)]
-pub struct TriggerDependencies(pub Vec<TriggerDependency>);
-
-/// Component reference or inline definition
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
-#[serde(deny_unknown_fields, untagged, try_from = "toml::Value")]
-#[schemars(schema_with = "json_schema::id_or_component")]
-pub enum ComponentSpec {
-    /// `"component-id"`
-    Reference(KebabId),
-    /// `{ ... }`
-    Inline(Box<Component>),
-}
-
-/// Specifies how to satisfy an import dependency of the component. This may be one of:
-///
-/// - A semantic versioning constraint for the package version to use. Spin fetches the latest matching version of the package whose name matches the dependency name from the default registry.
-///
-/// Example: `"my:dep/import" = ">= 0.1.0"`
-///
-/// - A package from a registry.
-///
-/// Example: `"my:dep/import" = { version = "0.1.0", registry = "registry.io", ...}`
-///
-/// - A package from a filesystem path.
-///
-/// Example: `"my:dependency" = { path = "path/to/component.wasm", export = "my-export" }`
-///
-/// - A component in the application. The referenced component binary is composed: additional
-///   configuration such as files, networking, storage, etc. are ignored. This is intended
-///   primarily as a convenience for including dependencies in the manifest so that they
-///   can be built using `spin build`.
-///
-/// Example: `"my:dependency" = { component = "my-dependency", export = "my-export" }`
-///
-/// - A package from an HTTP URL.
-///
-/// Example: `"my:import" = { url = "https://example.com/component.wasm", sha256 = "sha256:..." }`
-///
-/// Learn more: https://spinframework.dev/v3/writing-apps#using-component-dependencies
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
-#[serde(untagged, deny_unknown_fields)]
-pub enum TriggerDependency {
-    /// `... = { version = "0.1.0", registry = "registry.io", ...}`
-    #[schemars(description = "")] // schema docs are on the parent
-    Package {
-        /// A semantic versioning constraint for the package version to use. Required. Spin
-        /// fetches the latest matching version from the specified registry, or from
-        /// the default registry if no registry is specified.
-        ///
-        /// Example: `"my:dep/import" = { version = ">= 0.1.0" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-registry
-        version: String,
-        /// The registry that hosts the package. If omitted, this defaults to your
-        /// system default registry.
-        ///
-        /// Example: `"my:dep/import" = { registry = "registry.io", version = "0.1.0" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-registry
-        registry: Option<String>,
-        /// The name of the package to use. If omitted, this defaults to the package name of the
-        /// imported interface.
-        ///
-        /// Example: `"my:dep/import" = { package = "your:implementation", version = "0.1.0" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-registry
-        package: String,
-        /// The set of configurations to inherit from the parent component. If omitted or set to `false`,
-        /// no configurations will be inherited. If `true`, all configurations will be inherited.
-        /// Selective inheritance can be specified by enumerating the configuration keys the dependency
-        /// would like to inherit.
-        ///
-        /// Examples:
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = true }`
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = ["ai_models", "allowed_outbound_hosts"] }`
-        inherit_configuration: Option<InheritConfiguration>,
-    },
-    /// `... = { path = "path/to/component.wasm", export = "my-export" }`
-    #[schemars(description = "")] // schema docs are on the parent
-    Local {
-        /// The path to the Wasm file that implements the dependency.
-        ///
-        /// Example: `"my:dep/import" = { path = "path/to/component.wasm" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-local-component
-        path: PathBuf,
-        /// The set of configurations to inherit from the parent component. If omitted or set to `false`,
-        /// no configurations will be inherited. If `true`, all configurations will be inherited.
-        /// Selective inheritance can be specified by enumerating the configuration keys the dependency
-        /// would like to inherit.
-        ///
-        /// Examples:
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = true }`
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = ["ai_models", "allowed_outbound_hosts"] }`
-        inherit_configuration: Option<InheritConfiguration>,
-    },
-    /// `... = { url = "https://example.com/component.wasm", sha256 = "..." }`
-    #[schemars(description = "")] // schema docs are on the parent
-    HTTP {
-        /// The URL to the Wasm component that implements the dependency.
-        ///
-        /// Example: `"my:dep/import" = { url = "https://example.com/component.wasm", sha256 = "sha256:..." }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-url
-        url: String,
-        /// The SHA256 digest of the Wasm file. This is required for integrity checking. Must begin with `sha256:`.
-        ///
-        /// Example: `"my:dep/import" = { sha256 = "sha256:...", ... }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-url
-        digest: String,
-        /// The set of configurations to inherit from the parent component. If omitted or set to `false`,
-        /// no configurations will be inherited. If `true`, all configurations will be inherited.
-        /// Selective inheritance can be specified by enumerating the configuration keys the dependency
-        /// would like to inherit.
-        ///
-        /// Examples:
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = true }`
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = ["ai_models", "allowed_outbound_hosts"] }`
-        inherit_configuration: Option<InheritConfiguration>,
-    },
-    /// `... = { component = "my-dependency" }`
-    #[schemars(description = "")] // schema docs are on the parent
-    AppComponent {
-        /// The ID of the component which implements the dependency.
-        ///
-        /// Example: `"my:dep/import" = { component = "my-dependency" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#using-component-dependencies
-        component: KebabId,
-        /// The set of configurations to inherit from the parent component. If omitted or set to `false`,
-        /// no configurations will be inherited. If `true`, all configurations will be inherited.
-        /// Selective inheritance can be specified by enumerating the configuration keys the dependency
-        /// would like to inherit.
-        ///
-        /// Examples:
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = true }`
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = ["ai_models", "allowed_outbound_hosts"] }`
-        inherit_configuration: Option<InheritConfiguration>,
-    },
-}
-
-impl TryFrom<toml::Value> for ComponentSpec {
-    type Error = toml::de::Error;
-
-    fn try_from(value: toml::Value) -> Result<Self, Self::Error> {
-        if value.is_str() {
-            Ok(ComponentSpec::Reference(KebabId::deserialize(value)?))
-        } else {
-            Ok(ComponentSpec::Inline(Box::new(Component::deserialize(
-                value,
-            )?)))
-        }
-    }
-}
-
-/// Specifies how to satisfy an import dependency of the component. This may be one of:
-///
-/// - A semantic versioning constraint for the package version to use. Spin fetches the latest matching version of the package whose name matches the dependency name from the default registry.
-///
-/// Example: `"my:dep/import" = ">= 0.1.0"`
-///
-/// - A package from a registry.
-///
-/// Example: `"my:dep/import" = { version = "0.1.0", registry = "registry.io", ...}`
-///
-/// - A package from a filesystem path.
-///
-/// Example: `"my:dependency" = { path = "path/to/component.wasm", export = "my-export" }`
-///
-/// - A component in the application. The referenced component binary is composed: additional
-///   configuration such as files, networking, storage, etc. are ignored. This is intended
-///   primarily as a convenience for including dependencies in the manifest so that they
-///   can be built using `spin build`.
-///
-/// Example: `"my:dependency" = { component = "my-dependency", export = "my-export" }`
-///
-/// - A package from an HTTP URL.
-///
-/// Example: `"my:import" = { url = "https://example.com/component.wasm", sha256 = "sha256:..." }`
-///
-/// Learn more: https://spinframework.dev/v3/writing-apps#using-component-dependencies
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
-#[serde(untagged, deny_unknown_fields)]
-pub enum ComponentDependency {
-    /// `... = ">= 0.1.0"`
-    #[schemars(description = "")] // schema docs are on the parent
-    Version(String),
-    /// `... = { version = "0.1.0", registry = "registry.io", ...}`
-    #[schemars(description = "")] // schema docs are on the parent
-    Package {
-        /// A semantic versioning constraint for the package version to use. Required. Spin
-        /// fetches the latest matching version from the specified registry, or from
-        /// the default registry if no registry is specified.
-        ///
-        /// Example: `"my:dep/import" = { version = ">= 0.1.0" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-registry
-        version: String,
-        /// The registry that hosts the package. If omitted, this defaults to your
-        /// system default registry.
-        ///
-        /// Example: `"my:dep/import" = { registry = "registry.io", version = "0.1.0" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-registry
-        registry: Option<String>,
-        /// The name of the package to use. If omitted, this defaults to the package name of the
-        /// imported interface.
-        ///
-        /// Example: `"my:dep/import" = { package = "your:implementation", version = "0.1.0" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-registry
-        package: Option<String>,
-        /// The name of the export in the package. If omitted, this defaults to the name of the import.
-        ///
-        /// Example: `"my:dep/import" = { export = "your:impl/export", version = "0.1.0" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-registry
-        export: Option<String>,
-        /// The set of configurations to inherit from the parent component. If omitted or set to `false`,
-        /// no configurations will be inherited. If `true`, all configurations will be inherited.
-        /// Selective inheritance can be specified by enumerating the configuration keys the dependency
-        /// would like to inherit.
-        ///
-        /// Examples:
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = true }`
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = ["ai_models", "allowed_outbound_hosts"] }`
-        inherit_configuration: Option<InheritConfiguration>,
-    },
-    /// `... = { path = "path/to/component.wasm", export = "my-export" }`
-    #[schemars(description = "")] // schema docs are on the parent
-    Local {
-        /// The path to the Wasm file that implements the dependency.
-        ///
-        /// Example: `"my:dep/import" = { path = "path/to/component.wasm" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-local-component
-        path: PathBuf,
-        /// The name of the export in the package. If omitted, this defaults to the name of the import.
-        ///
-        /// Example: `"my:dep/import" = { export = "your:impl/export", path = "path/to/component.wasm" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-local-component
-        export: Option<String>,
-        /// The set of configurations to inherit from the parent component. If omitted or set to `false`,
-        /// no configurations will be inherited. If `true`, all configurations will be inherited.
-        /// Selective inheritance can be specified by enumerating the configuration keys the dependency
-        /// would like to inherit.
-        ///
-        /// Examples:
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = true }`
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = ["ai_models", "allowed_outbound_hosts"] }`
-        inherit_configuration: Option<InheritConfiguration>,
-    },
-    /// `... = { url = "https://example.com/component.wasm", sha256 = "..." }`
-    #[schemars(description = "")] // schema docs are on the parent
-    HTTP {
-        /// The URL to the Wasm component that implements the dependency.
-        ///
-        /// Example: `"my:dep/import" = { url = "https://example.com/component.wasm", sha256 = "sha256:..." }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-url
-        url: String,
-        /// The SHA256 digest of the Wasm file. This is required for integrity checking. Must begin with `sha256:`.
-        ///
-        /// Example: `"my:dep/import" = { sha256 = "sha256:...", ... }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-url
-        digest: String,
-        /// The name of the export in the package. If omitted, this defaults to the name of the import.
-        ///
-        /// Example: `"my:dep/import" = { export = "your:impl/export", ... }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#dependencies-from-a-url
-        export: Option<String>,
-        /// The set of configurations to inherit from the parent component. If omitted or set to `false`,
-        /// no configurations will be inherited. If `true`, all configurations will be inherited.
-        /// Selective inheritance can be specified by enumerating the configuration keys the dependency
-        /// would like to inherit.
-        ///
-        /// Examples:
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = true }`
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = ["ai_models", "allowed_outbound_hosts"] }`
-        inherit_configuration: Option<InheritConfiguration>,
-    },
-    /// `... = { component = "my-dependency" }`
-    #[schemars(description = "")] // schema docs are on the parent
-    AppComponent {
-        /// The ID of the component which implements the dependency.
-        ///
-        /// Example: `"my:dep/import" = { component = "my-dependency" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#using-component-dependencies
-        component: KebabId,
-        /// The name of the export in the package. If omitted, this defaults to the name of the import.
-        ///
-        /// Example: `"my:dep/import" = { export = "your:impl/export", component = "my-dependency" }`
-        ///
-        /// Learn more: https://spinframework.dev/writing-apps#using-component-dependencies
-        export: Option<String>,
-        /// The set of configurations to inherit from the parent component. If omitted or set to `false`,
-        /// no configurations will be inherited. If `true`, all configurations will be inherited.
-        /// Selective inheritance can be specified by enumerating the configuration keys the dependency
-        /// would like to inherit.
-        ///
-        /// Examples:
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = true }`
-        ///     `"my:dep/import" = { version = "0.1.0", inherit_configuration = ["ai_models", "allowed_outbound_hosts"] }`
-        inherit_configuration: Option<InheritConfiguration>,
-    },
-}
-
-/// The set of configurations to inherit from the parent component.
-///
-/// Can be specified as:
-/// - `true` — inherit all configurations
-/// - `false` — inherit no configurations (equivalent to omitting the field)
-/// - `["key1", "key2"]` — inherit only the specified configuration keys
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
-#[serde(untagged)]
-pub enum InheritConfiguration {
-    /// All or no configurations will be inherited, specified as `true` or `false`.
-    All(bool),
-    /// Only the specified configuration keys will be inherited from the parent component.
-    Some(Vec<String>),
-}
-
-impl ComponentDependency {
-    /// Returns the `inherit_configuration` field if present on this dependency variant.
-    pub fn inherit_configuration(&self) -> Option<&InheritConfiguration> {
-        match self {
-            ComponentDependency::Version(_) => None,
-            ComponentDependency::Package {
-                inherit_configuration,
-                ..
-            }
-            | ComponentDependency::Local {
-                inherit_configuration,
-                ..
-            }
-            | ComponentDependency::HTTP {
-                inherit_configuration,
-                ..
-            }
-            | ComponentDependency::AppComponent {
-                inherit_configuration,
-                ..
-            } => inherit_configuration.as_ref(),
-        }
-    }
-
-    /// Sets the `inherit_configuration` field on this dependency variant.
-    /// No-op for `Version` variants.
-    pub fn set_inherit_configuration(&mut self, value: InheritConfiguration) {
-        match self {
-            ComponentDependency::Version(_) => {}
-            ComponentDependency::Package {
-                inherit_configuration,
-                ..
-            }
-            | ComponentDependency::Local {
-                inherit_configuration,
-                ..
-            }
-            | ComponentDependency::HTTP {
-                inherit_configuration,
-                ..
-            }
-            | ComponentDependency::AppComponent {
-                inherit_configuration,
-                ..
-            } => {
-                *inherit_configuration = Some(value);
-            }
-        }
-    }
-}
-
-/// A Spin component.
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
-#[serde(deny_unknown_fields)]
-pub struct Component {
-    /// The file, package, or URL containing the component Wasm binary.
-    ///
-    /// Example: `source = "bin/cart.wasm"`
-    ///
-    /// Learn more: https://spinframework.dev/writing-apps#the-component-source
-    pub source: ComponentSource,
-    /// A human-readable description of the component.
-    ///
-    /// Example: `description = "Shopping cart"`
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub description: String,
-    /// Configuration variables available to the component. Names must be
-    /// in `lower_snake_case`. Values are strings, and may refer
-    /// to application variables using `{{ ... }}` syntax.
-    ///
-    /// `variables = { users_endpoint = "https://{{ api_host }}/users"}`
-    ///
-    /// Learn more: https://spinframework.dev/variables#adding-variables-to-your-applications
-    #[serde(default, skip_serializing_if = "Map::is_empty")]
-    pub variables: Map<LowerSnakeId, String>,
-    /// Environment variables to be set for the Wasm module.
-    ///
-    /// `environment = { DB_URL = "mysql://spin:spin@localhost/dev" }`
-    #[serde(default, skip_serializing_if = "Map::is_empty")]
-    pub environment: Map<String, String>,
-    /// The files the component is allowed to read. Each list entry is either:
-    ///
-    /// - a glob pattern (e.g. "assets/**/*.jpg"); or
-    ///
-    /// - a source-destination pair indicating where a host directory should be mapped in the guest (e.g. { source = "assets", destination = "/" })
-    ///
-    /// Learn more: https://spinframework.dev/writing-apps#including-files-with-components
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub files: Vec<WasiFilesMount>,
-    /// Any files or glob patterns that should not be available to the
-    /// Wasm module at runtime, even though they match a `files`` entry.
-    ///
-    /// Example: `exclude_files = ["secrets/*"]`
-    ///
-    /// Learn more: https://spinframework.dev/writing-apps#including-files-with-components
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub exclude_files: Vec<String>,
-    /// Deprecated. Use `allowed_outbound_hosts` instead.
-    ///
-    /// Example: `allowed_http_hosts = ["example.com"]`
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    #[deprecated]
-    pub allowed_http_hosts: Vec<String>,
-    /// The network destinations which the component is allowed to access.
-    /// Each entry is in the form "(scheme)://(host)[:port]". Each element
-    /// allows * as a wildcard e.g. "https://\*" (HTTPS on the default port
-    /// to any destination) or "\*://localhost:\*" (any protocol to any port on
-    /// localhost). The host part allows segment wildcards for subdomains
-    /// e.g. "https://\*.example.com". Application variables are allowed using
-    /// `{{ my_var }}`` syntax.
-    ///
-    /// Example: `allowed_outbound_hosts = ["redis://myredishost.com:6379"]`
-    ///
-    /// Learn more: https://spinframework.dev/http-outbound#granting-http-permissions-to-components
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    #[schemars(with = "Vec<json_schema::AllowedOutboundHost>")]
-    pub allowed_outbound_hosts: Vec<String>,
-    /// The key-value stores which the component is allowed to access. Stores are identified
-    /// by label e.g. "default" or "customer". Stores other than "default" must be mapped
-    /// to a backing store in the runtime config.
-    ///
-    /// Example: `key_value_stores = ["default", "my-store"]`
-    ///
-    /// Learn more: https://spinframework.dev/kv-store-api-guide#custom-key-value-stores
-    #[serde(
-        default,
-        with = "kebab_or_snake_case",
-        skip_serializing_if = "Vec::is_empty"
-    )]
-    #[schemars(with = "Vec<json_schema::KeyValueStore>")]
-    pub key_value_stores: Vec<String>,
-    /// The SQLite databases which the component is allowed to access. Databases are identified
-    /// by label e.g. "default" or "analytics". Databases other than "default" must be mapped
-    /// to a backing store in the runtime config. Use "spin up --sqlite" to run database setup scripts.
-    ///
-    /// Example: `sqlite_databases = ["default", "my-database"]`
-    ///
-    /// Learn more: https://spinframework.dev/sqlite-api-guide#preparing-an-sqlite-database
-    #[serde(
-        default,
-        with = "kebab_or_snake_case",
-        skip_serializing_if = "Vec::is_empty"
-    )]
-    #[schemars(with = "Vec<json_schema::SqliteDatabase>")]
-    pub sqlite_databases: Vec<String>,
-    /// The AI models which the component is allowed to access. For local execution, you must
-    /// download all models; for hosted execution, you should check which models are available
-    /// in your target environment.
-    ///
-    /// Example: `ai_models = ["llama2-chat"]`
-    ///
-    /// Learn more: https://spinframework.dev/serverless-ai-api-guide#using-serverless-ai-from-applications
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    #[schemars(with = "Vec<json_schema::AIModel>")]
-    pub ai_models: Vec<String>,
-    /// The Spin environments with which the component must be compatible.
-    /// If present, this overrides the default application targets (they are not combined).
-    ///
-    /// Example: `targets = ["spin-up:3.3", "spinkube:0.4"]`
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub targets: Option<Vec<TargetEnvironmentRef>>,
-    /// The component build configuration.
-    ///
-    /// Learn more: https://spinframework.dev/build
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub build: Option<ComponentBuildConfig>,
-    /// Settings for custom tools or plugins. Spin ignores this field.
-    #[serde(default, skip_serializing_if = "Map::is_empty")]
-    #[schemars(schema_with = "json_schema::map_of_toml_tables")]
-    pub tool: Map<String, toml::Table>,
-    /// If true, dependencies can invoke Spin APIs with the same permissions as the main
-    /// component. If false, dependencies have no permissions (e.g. network,
-    /// key-value stores, SQLite databases).
-    ///
-    /// Learn more: https://spinframework.dev/writing-apps#dependency-permissions
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub dependencies_inherit_configuration: Option<bool>,
-    /// Specifies how to satisfy Wasm Component Model imports of this component.
-    ///
-    /// Learn more: https://spinframework.dev/writing-apps#using-component-dependencies
-    #[serde(default, skip_serializing_if = "ComponentDependencies::is_empty")]
-    pub dependencies: ComponentDependencies,
-    /// Override values to use when building or running a named build profile.
-    ///
-    /// Example: `profile.debug.build.command = "npm run build-debug"`
-    #[serde(default, skip_serializing_if = "Map::is_empty")]
-    pub(crate) profile: Map<String, ComponentProfileOverride>,
-}
-
-/// Customisations for a Spin component in a non-default profile.
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
-#[serde(deny_unknown_fields)]
-pub struct ComponentProfileOverride {
-    /// The file, package, or URL containing the component Wasm binary.
-    ///
-    /// Example: `source = "bin/debug/cart.wasm"`
-    ///
-    /// Learn more: https://spinframework.dev/writing-apps#the-component-source
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub(crate) source: Option<ComponentSource>,
-
-    /// Environment variables for the Wasm module to be overridden in this profile.
-    /// Environment variables specified in the default profile will still be set
-    /// if not overridden here.
-    ///
-    /// `environment = { DB_URL = "mysql://spin:spin@localhost/dev" }`
-    #[serde(default, skip_serializing_if = "Map::is_empty")]
-    pub(crate) environment: Map<String, String>,
-
-    /// Wasm Component Model imports to be overridden in this profile.
-    /// Dependencies specified in the default profile will still be composed
-    /// if not overridden here.
-    ///
-    /// Learn more: https://spinframework.dev/writing-apps#using-component-dependencies
-    #[serde(default, skip_serializing_if = "ComponentDependencies::is_empty")]
-    pub(crate) dependencies: ComponentDependencies,
-
-    /// The command or commands for building the component in non-default profiles.
-    /// If a component has no special build instructions for a profile, the
-    /// default build command is used.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub(crate) build: Option<ComponentProfileBuildOverride>,
-}
-
-/// Customisations for a Spin component build in a non-default profile.
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
-#[serde(deny_unknown_fields)]
-pub struct ComponentProfileBuildOverride {
-    /// The command or commands to build the component in a named profile. If multiple commands
-    /// are specified, they are run sequentially from left to right.
-    ///
-    /// Example: `build.command = "cargo build"`
-    ///
-    /// Learn more: https://spinframework.dev/build#setting-up-for-spin-build
-    pub(crate) command: super::common::Commands,
-}
-
-/// Component dependencies
-#[derive(Clone, Debug, Default, Serialize, Deserialize, JsonSchema)]
-#[serde(transparent)]
-pub struct ComponentDependencies {
-    /// `dependencies = { "foo:bar" = ">= 0.1.0" }`
-    pub inner: Map<DependencyName, ComponentDependency>,
-}
-
-impl ComponentDependencies {
-    /// This method validates the correct specification of dependencies in a
-    /// component section of the manifest. See the documentation on the methods
-    /// called for more information on the specific checks.
-    fn validate(&self) -> anyhow::Result<()> {
-        self.ensure_plain_names_have_package()?;
-        self.ensure_package_names_no_export()?;
-        self.ensure_disjoint()?;
-        Ok(())
-    }
-
-    /// This method ensures that all dependency names in plain form (e.g.
-    /// "foo-bar") do not map to a `ComponentDependency::Version`, or a
-    /// `ComponentDependency::Package` where the `package` is `None`.
-    fn ensure_plain_names_have_package(&self) -> anyhow::Result<()> {
-        for (dependency_name, dependency) in self.inner.iter() {
-            let DependencyName::Plain(plain) = dependency_name else {
-                continue;
-            };
-            match dependency {
-                ComponentDependency::Package { package, .. } if package.is_none() => {}
-                ComponentDependency::Version(_) => {}
-                _ => continue,
-            }
-            anyhow::bail!("dependency {plain:?} must specify a package name");
-        }
-        Ok(())
-    }
-
-    /// This method ensures that dependency names in the package form (e.g.
-    /// "foo:bar" or "foo:bar@0.1.0") do not map to specific exported
-    /// interfaces, e.g. `"foo:bar = { ..., export = "my-export" }"` is invalid.
-    fn ensure_package_names_no_export(&self) -> anyhow::Result<()> {
-        for (dependency_name, dependency) in self.inner.iter() {
-            if let DependencyName::Package(name) = dependency_name
-                && name.interface.is_none()
-            {
-                let export = match dependency {
-                    ComponentDependency::Package { export, .. } => export,
-                    ComponentDependency::Local { export, .. } => export,
-                    _ => continue,
-                };
-
-                anyhow::ensure!(
-                    export.is_none(),
-                    "using an export to satisfy the package dependency {dependency_name:?} is not currently permitted",
-                );
-            }
-        }
-        Ok(())
-    }
-
-    /// This method ensures that dependencies names do not conflict with each other. That is to say
-    /// that two dependencies of the same package must have disjoint versions or interfaces.
-    fn ensure_disjoint(&self) -> anyhow::Result<()> {
-        for [this, other] in self.inner.keys().array_combinations::<2>() {
-            let (DependencyName::Package(this), DependencyName::Package(other)) = (this, other)
-            else {
-                continue;
-            };
-
-            if this.package == other.package {
-                Self::check_disjoint(this, other)?;
-            }
-        }
-        Ok(())
-    }
-
-    fn check_disjoint(
-        this: &DependencyPackageName,
-        other: &DependencyPackageName,
-    ) -> anyhow::Result<()> {
-        assert_eq!(this.package, other.package);
-
-        if let (Some(this_ver), Some(other_ver)) = (this.version.clone(), other.version.clone())
-            && Self::normalize_compatible_version(this_ver)
-                != Self::normalize_compatible_version(other_ver)
-        {
-            return Ok(());
-        }
-
-        if let (Some(this_itf), Some(other_itf)) =
-            (this.interface.as_ref(), other.interface.as_ref())
-            && this_itf != other_itf
-        {
-            return Ok(());
-        }
-
-        Err(anyhow!("{this:?} dependency conflicts with {other:?}"))
-    }
-
-    /// Normalize version to perform a compatibility check against another version.
-    ///
-    /// See backwards comptabilitiy rules at https://semver.org/
-    fn normalize_compatible_version(mut version: semver::Version) -> semver::Version {
-        version.build = semver::BuildMetadata::EMPTY;
-
-        if version.pre != semver::Prerelease::EMPTY {
-            return version;
-        }
-        if version.major > 0 {
-            version.minor = 0;
-            version.patch = 0;
-            return version;
-        }
-
-        if version.minor > 0 {
-            version.patch = 0;
-            return version;
-        }
-
-        version
-    }
-
-    fn is_empty(&self) -> bool {
-        self.inner.is_empty()
-    }
-}
-
-/// Identifies a deployment target.
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema)]
-#[serde(untagged, deny_unknown_fields)]
-pub enum TargetEnvironmentRef {
-    /// Environment definition doc reference e.g. `spin-up:3.2`, `my-host`. This is looked up
-    /// in the default environment catalogue (the `spin-environments` repo, `env` directory).
-    Catalogue(String),
-    /// An environment definition doc HTTP URL
-    Http {
-        /// The environment document URL e.g. `https://github.com/me/environments/blob/main/target-envs/spin-up.3.6.toml`.
-        url: String,
-    },
-    /// A local environment document file. This is expected to contain a serialised
-    /// EnvironmentDefinition in TOML format.
-    File {
-        /// The file path of the document.
-        path: PathBuf,
-    },
-}
-
-impl std::fmt::Display for TargetEnvironmentRef {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Catalogue(e) => e.fmt(f),
-            Self::Http { url } => url.fmt(f),
-            Self::File { path } => path.display().fmt(f),
-        }
-    }
-}
-
-mod kebab_or_snake_case {
-    use serde::{Deserialize, Serialize};
-    pub use spin_serde::{KebabId, SnakeId};
-    pub fn serialize<S>(value: &[String], serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::ser::Serializer,
-    {
-        if value.iter().all(|s| {
-            KebabId::try_from(s.clone()).is_ok() || SnakeId::try_from(s.to_owned()).is_ok()
-        }) {
-            value.serialize(serializer)
-        } else {
-            Err(serde::ser::Error::custom(
-                "expected kebab-case or snake_case",
-            ))
-        }
-    }
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<Vec<String>, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let value = toml::Value::deserialize(deserializer)?;
-        let list: Vec<String> = Vec::deserialize(value).map_err(serde::de::Error::custom)?;
-        if list.iter().all(|s| {
-            KebabId::try_from(s.clone()).is_ok() || SnakeId::try_from(s.to_owned()).is_ok()
-        }) {
-            Ok(list)
-        } else {
-            Err(serde::de::Error::custom(
-                "expected kebab-case or snake_case",
-            ))
-        }
-    }
-}
-
-impl Component {
-    /// Combine `allowed_outbound_hosts` with the deprecated `allowed_http_hosts` into
-    /// one array all normalized to the syntax of `allowed_outbound_hosts`.
-    pub fn normalized_allowed_outbound_hosts(&self) -> anyhow::Result<Vec<String>> {
-        #[allow(deprecated)]
-        let normalized =
-            crate::compat::convert_allowed_http_to_allowed_hosts(&self.allowed_http_hosts, false)?;
-        if !normalized.is_empty() {
-            terminal::warn!(
-                "Use of the deprecated field `allowed_http_hosts` - to fix, \
-            replace `allowed_http_hosts` with `allowed_outbound_hosts = {normalized:?}`",
-            )
-        }
-
-        Ok(self
-            .allowed_outbound_hosts
-            .iter()
-            .cloned()
-            .chain(normalized)
-            .collect())
-    }
-}
-
-mod one_or_many {
-    use serde::{Deserialize, Deserializer, Serialize, Serializer};
-
-    pub fn serialize<T, S>(vec: &Vec<T>, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        T: Serialize,
-        S: Serializer,
-    {
-        if vec.len() == 1 {
-            vec[0].serialize(serializer)
-        } else {
-            vec.serialize(serializer)
-        }
-    }
-
-    pub fn deserialize<'de, T, D>(deserializer: D) -> Result<Vec<T>, D::Error>
-    where
-        T: Deserialize<'de>,
-        D: Deserializer<'de>,
-    {
-        let value = toml::Value::deserialize(deserializer)?;
-        // NOTE: We explicitly check for array first rather than trying T::deserialize
-        // first, because toml's serde impl will treat an array as a sequence of fields
-        // to be assigned to struct members (e.g. Component), producing nonsensical results.
-        if let Some(arr) = value.as_array() {
-            arr.iter()
-                .map(|v| T::deserialize(v.clone()))
-                .collect::<Result<Vec<_>, _>>()
-                .map_err(serde::de::Error::custom)
-        } else {
-            T::deserialize(value)
-                .map(|v| vec![v])
-                .map_err(serde::de::Error::custom)
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
+    use std::path::PathBuf;
     use toml::toml;
 
     use super::*;
